requirements.txt
qiskit-terra>=0.16.0
qiskit-ignis>=0.4.0
scipy>=1.4
sympy>=1.3
numpy>=1.17
psutil>=5
scikit-learn>=0.20.0
dlx
docplex
fastdtw
setuptools>=40.1.0
h5py
pandas
quandl
yfinance
retworkx>=0.5.0
python-socketio
